fix GroupDelete

+++ src/main/java/mops/gruppen2/controller/Gruppen2Controller.java
@@ -43,15 +43,15 @@ public class Gruppen2Controller {
     private final UserService userService;
     private final ControllerService controllerService;
     private final InviteLinkRepositoryService inviteLinkRepositoryService;
-    @Autowired
-    Gruppen2Config gruppen2Config;
+    private final Gruppen2Config gruppen2Config;
 
-    public Gruppen2Controller(KeyCloakService keyCloakService, GroupService groupService, UserService userService, ControllerService controllerService, InviteLinkRepositoryService inviteLinkRepositoryService) {
+    public Gruppen2Controller(KeyCloakService keyCloakService, GroupService groupService, UserService userService, ControllerService controllerService, InviteLinkRepositoryService inviteLinkRepositoryService, Gruppen2Config gruppen2Config) {
         this.keyCloakService = keyCloakService;
         this.groupService = groupService;
         this.userService = userService;
         this.controllerService = controllerService;
         this.inviteLinkRepositoryService = inviteLinkRepositoryService;
+        this.gruppen2Config = gruppen2Config;
     }
 
     /**
@@ -258,6 +258,10 @@ public class Gruppen2Controller {
     public String pDeleteGroup(KeycloakAuthenticationToken token, @RequestParam("group_id") Long groupId){
         Account account = keyCloakService.createAccountFromPrincipal(token);
         User user = new User(account.getName(), account.getGivenname(), account.getFamilyname(), account.getEmail());
+        Group group = userService.getGroupById(groupId);
+        if(group.getRoles().get(user.getId()) != Role.ADMIN ){
+            return "error";
+        }
         controllerService.deleteGroupEvent(user.getId(), groupId);
         return "redirect:/gruppen2/";
     }
